layers/serving_layer/requirements.txt
# FastAPI Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
pydantic==2.9.2
pydantic-settings==2.6.1

# MongoDB
pymongo==4.9.1
motor==3.6.0  # Async MongoDB driver

# Redis
redis==5.2.0
hiredis==3.0.0  # C parser for better performance

# Data processing
python-dateutil==2.9.0

# YAML configuration
PyYAML==6.0.2

# Logging
python-json-logger==3.2.1

# Authentication
PyJWT==2.9.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2  # For testing FastAPI

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
